spring-mvc-final-feature-JBGW06-028/nhnpage/src/test/java/com/nhnacademy/nhnpage/LogoutControllerTest.java
package com.nhnacademy.nhnpage;

import com.nhnacademy.nhnpage.controller.LogoutController;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.mock.web.MockHttpSession;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.context.WebApplicationContext;

import jakarta.servlet.http.Cookie;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(LogoutController.class)
public class LogoutControllerTest {
    private MockMvc mockMvc;

    @BeforeEach
    void setUp(WebApplicationContext wac) {
        this.mockMvc = MockMvcBuilders.webAppContextSetup(wac).build();
    }

    @Test
    void testLogout() throws Exception {
        MockHttpSession session = new MockHttpSession();
        session.setAttribute("id", "user");

        Cookie sessionCookie = new Cookie("SESSIONNHN", session.getId());

        mockMvc.perform(get("/cs/logout")
                        .session(session)
                        .cookie(sessionCookie))
                .andExpect(status().is3xxRedirection())
                .andExpect(redirectedUrl("/cs/login"))
                .andExpect(request().sessionAttributeDoesNotExist("id"));

        // Ensure session is invalidated
        assert session.isInvalid();
    }
}
